requirements.txt
flask
groq
PyPDF2
python-docx
python-dotenv
